Allow for use of opaque AES KEKs
The output from the HKDF step in KEK generation is currently assumed to
result in explicit key material, which is then used with the `Aes` trait
implementation to perform key encryption/decryption.

Allow for the possibility of this AES key staying in an opaque form by
using new `.._aes()` variants of the HKDF trait methods, which emit
`OpaqueOr<aes::Key>` results of the specified size.

The default implementations of these new methods perform the existing
behaviour (generating explicit AES key material), so no changes are
needed in existing KeyMint codebases.

Test: VtsAidlKeyMintTargetTest
Bug: 373463166
Change-Id: I9200417724c6031d072988d8a5dcd0a6348f19eb